Update base_model.py
it  checks if 'id', 'created_at', and 'updated_at' attributes are present in the kwargs dictionary. If not, it assigns default values to them using uuid.uuid4() for 'id' and datetime.now() for 'created_at' and 'updated_at'.
It then iterates through the remaining kwargs items (excluding __class__) and sets the corresponding attributes of the instance. it  alsochecks if 'created_at' or 'updated_at' is present in kwargs and, if so, converts the string representation of datetime to a datetime object